legacy-shit/requirements.txt
pyyaml
spore
asyncio
encodium